frontend/requirements.txt
backports.functools-lru-cache==1.5
boto3==1.9.74
botocore==1.12.74
certifi==2018.11.29
chardet==3.0.4
cycler==0.10.0
docutils==0.14
futures==3.2.0
idna==2.8
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.3
numpy==1.15.4
Pillow==5.4.1
pyparsing==2.3.0
python-dateutil==2.7.5
pytz==2018.9
requests==2.21.0
s3transfer==0.1.13
six==1.12.0
subprocess32==3.5.3
urllib3==1.24.1
